requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
chromadb>=0.4.22
pymupdf>=1.23.0
python-docx>=1.1.0
openai>=1.12.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
sentence-transformers>=2.4.0